refactor some functions
